+++ oap-server/mqe-rt/src/main/java/org/apache/skywalking/mqe/rt/operation/CompareOp.java
@@ -20,7 +20,7 @@ package org.apache.skywalking.mqe.rt.operation;
 
 import org.apache.skywalking.mqe.rt.exception.IllegalExpressionException;
 import org.apache.skywalking.mqe.rt.grammar.MQEParser;
-import org.apache.skywalking.mqe.rt.type.ExpressionResult;
+import org.apache.skywalking.oap.server.core.query.mqe.ExpressionResult;
 
 public class CompareOp {
     public static ExpressionResult doCompareOP(ExpressionResult left,