+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/query/IAlarmQueryDAO.java
@@ -18,6 +18,7 @@
 
 package org.apache.skywalking.oap.server.core.storage.query;
 
+import com.google.gson.JsonObject;
 import java.io.IOException;
 import java.util.Base64;
 import java.util.List;
@@ -25,11 +26,19 @@ import java.util.List;
 import com.google.common.base.Charsets;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
+import org.apache.skywalking.oap.server.core.alarm.AlarmRecord;
+import org.apache.skywalking.oap.server.core.alarm.AlarmSnapshotRecord;
 import org.apache.skywalking.oap.server.core.analysis.manual.searchtag.Tag;
+import org.apache.skywalking.oap.server.core.query.enumeration.Scope;
 import org.apache.skywalking.oap.server.core.query.input.Duration;
+import org.apache.skywalking.oap.server.core.query.mqe.MQEMetric;
+import org.apache.skywalking.oap.server.core.query.mqe.MQEValues;
+import org.apache.skywalking.oap.server.core.query.type.AlarmMessage;
+import org.apache.skywalking.oap.server.core.query.type.AlarmSnapshot;
 import org.apache.skywalking.oap.server.core.query.type.Alarms;
 import org.apache.skywalking.oap.server.core.query.type.KeyValue;
 import org.apache.skywalking.oap.server.core.storage.DAO;
+import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 public interface IAlarmQueryDAO extends DAO {
 
@@ -39,18 +48,54 @@ public interface IAlarmQueryDAO extends DAO {
                     final Duration duration, final List<Tag> tags) throws IOException;
 
     /**
-     * Parser the raw tags.
+     * Parse the raw tags.
      */
-    default void parserDataBinaryBase64(String dataBinaryBase64, List<KeyValue> tags) {
-        parserDataBinary(Base64.getDecoder().decode(dataBinaryBase64), tags);
+    default void parseDataBinaryBase64(String dataBinaryBase64, List<KeyValue> tags) {
+        parseDataBinary(Base64.getDecoder().decode(dataBinaryBase64), tags);
     }
 
     /**
-     * Parser the raw tags.
+     * Parse the raw tags.
      */
-    default void parserDataBinary(byte[] dataBinary, List<KeyValue> tags) {
+    default void parseDataBinary(byte[] dataBinary, List<KeyValue> tags) {
         List<Tag> tagList = GSON.fromJson(new String(dataBinary, Charsets.UTF_8), new TypeToken<List<Tag>>() {
         }.getType());
         tagList.forEach(pair -> tags.add(new KeyValue(pair.getKey(), pair.getValue())));
+    }
+
+    /**
+     * Build the alarm message from the alarm record.
+     * The Tags in JDBC storage is base64 encoded, need to decode in different way.
+     */
+    default AlarmMessage buildAlarmMessage(AlarmRecord alarmRecord) {
+        AlarmMessage message = new AlarmMessage();
+        message.setId(String.valueOf(alarmRecord.getId0()));
+        message.setId1(String.valueOf(alarmRecord.getId1()));
+        message.setName(alarmRecord.getName());
+        message.setMessage(alarmRecord.getAlarmMessage());
+        message.setStartTime(alarmRecord.getStartTime());
+        message.setScope(Scope.Finder.valueOf(alarmRecord.getScope()));
+        message.setScopeId(alarmRecord.getScope());
+        AlarmSnapshot alarmSnapshot = message.getSnapshot();
+        message.setSnapshot(alarmSnapshot);
+        String snapshot = alarmRecord.getSnapshot();
+        if (StringUtil.isNotBlank(snapshot)) {
+            AlarmSnapshotRecord alarmSnapshotRecord = GSON.fromJson(snapshot, AlarmSnapshotRecord.class);
+            alarmSnapshot.setExpression(alarmSnapshotRecord.getExpression());
+            JsonObject jsonObject = alarmSnapshotRecord.getMetrics();
+            if (jsonObject != null) {
+                for (final var obj : jsonObject.entrySet()) {
+                    final var name = obj.getKey();
+                    MQEMetric metrics = new MQEMetric();
+                    metrics.setName(name);
+                    List<MQEValues> values = GSON.fromJson(
+                        obj.getValue().getAsString(), new TypeToken<List<MQEValues>>() {
+                        }.getType());
+                    metrics.setResults(values);
+                    alarmSnapshot.getMetrics().add(metrics);
+                }
+            }
+        }
+        return message;
     }
 }