+++ docs/en/changes/changes.md
@@ -44,6 +44,8 @@
 * Add protection for dynamic config change propagate chain.
 * Add Ruby component IDs.
 * Add component ID(160) for Caffeine.
+* Alarm: Support store and query the metrics snapshot when the alarm is triggered.
+* Alarm: Remove unused `Alarm Trend` query.
 
 #### UI